Point to Slack `#pkg-registration` for manual merges
As [discussed on Slack earlier today](
https://julialang.slack.com/archives/C6M4DQA5P/p1677337269883019?thread_ts=1677284631.873199&cid=C6M4DQA5P),
and in issue #491, manual merges of new packages will usually only
happen after someone pings the `#pkg-registration` channel on Slack. For
new package developers who may not be aware of the Julia Slack, this may
cause considerable frustration, potentially alienating them from the
Julia ecosystem. Until there is some other workflow in place to handle
this entirely within GitHub, it seems best to explicitly recommend that
people post a message on Slack to get help with the manual merge.

This PR amends the instructions in the "If you do not want to fix the
AutoMerge issues..." bot message, asking the author to post to
`#pgk-registration` in addition to commenting on the GitHub issue.

Closes #491